requirements.txt
# Direct Production Dependencies
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.1
Pint==0.18
pymongo==4.0.1
scipy==1.7.3

# Direct Development Dependencies
mypy==0.931
mypy-extensions==0.4.3
pylint==2.12.2
pytest==6.2.5
pytest-cov==3.0.0
Sphinx==4.4.0
sphinx-rtd-theme==1.0.0

# Indirect Dependencies
alabaster==0.7.12
astroid==2.9.3
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.10
coverage==6.2
cycler==0.11.0
docutils==0.17.1
fonttools==4.29.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.10.1
iniconfig==1.1.1
isort==5.10.1
Jinja2==3.0.3
kafka-python
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
Pillow==9.0.0
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
snowballstemmer==2.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
